+++ api/src/controllers/zammood.js
@@ -95,12 +95,17 @@ router.post("/tickets", passport.authenticate(["referent", "young"], { session: 
   }
 });
 
-router.post("/ticketscount", passport.authenticate(["referent", "young"], { session: false, failWithError: true }), async (req, res) => {
+router.get("/ticketscount", passport.authenticate("referent", { session: false, failWithError: true }), async (req, res) => {
   try {
+    const user = await ReferentObject.findById(req.user._id);
+    let query = {};
+    if (user.role === ROLES.REFERENT_DEPARTMENT) query = { department: user.department, subject: "J'ai une question", role: "young", canal: "PLATFORM" };
+    if (user.role === ROLES.REFERENT_REGION) query = { region: user.region, subject: "J'ai une question", role: "young", canal: "PLATFORM" };
+
     const { ok, data } = await zammood.api(`/v0/ticket/count`, {
       method: "POST",
       credentials: "include",
-      body: JSON.stringify(req.body),
+      body: JSON.stringify(query),
     });
     if (!ok) return res.status(404).send({ ok: false, code: ERRORS.NOT_FOUND });