+++ CHANGELOG.md
@@ -1,3 +1,18 @@
+## [1.320.1](https://github.com/betagouv/service-national-universel/compare/v1.320.0...v1.320.1) (2023-08-09)
+
+
+### Bug Fixes
+
+* **admin, api:** tickets count ([#2865](https://github.com/betagouv/service-national-universel/issues/2865)) ([3305ac2](https://github.com/betagouv/service-national-universel/commit/3305ac25cc68c860716d7545a74fc05aef899039))
+* **admin/api:** es count by status ([#2867](https://github.com/betagouv/service-national-universel/issues/2867)) ([ff4020f](https://github.com/betagouv/service-national-universel/commit/ff4020f1e10aec2695b381c8cf2e0528bed5a1d4))
+
+# [1.320.0](https://github.com/betagouv/service-national-universel/compare/v1.319.0...v1.320.0) (2023-08-08)
+
+
+### Features
+
+* **lib:** update school level and inscriptionEndDate ([#2861](https://github.com/betagouv/service-national-universel/issues/2861)) ([2c51b7c](https://github.com/betagouv/service-national-universel/commit/2c51b7c238107a45e15a932e2b880affde24b6f1))
+
 # [1.319.0](https://github.com/betagouv/service-national-universel/compare/v1.318.2...v1.319.0) (2023-08-07)